+++ src/linted/assets.h
@@ -21,7 +21,7 @@
 typedef GLfloat linted_assets_point[2];
 linted_assets_point const *const linted_assets_triangle_data;
 
-GLuint const *const linted_assets_triangle_indices;
+GLubyte const *const linted_assets_triangle_indices;
 size_t const linted_assets_triangle_indices_size;
 
 #endif                          /* LINTED_ASSETS_H */
